fix methods, add tests, bump deps
